Convert JUnit tests to Kotlin

+++ src/main/java/com/weberapps/rayTracer/Tuple.kt
@@ -5,7 +5,7 @@ import java.lang.Math.abs
 import kotlin.math.sqrt
 
 const val EPSILON: Float = 0.00001f
-const val TAU: Float = (PI * 2).toFloat()
+const val TAU: Double = (PI * 2)
 open class Tuple(val x: Float, val y: Float, val z: Float, val w: Float) {
     operator fun plus(other: Tuple): Tuple {
         return Tuple(x + other.x, y + other.y, z + other.z, w + other.w)
